// apps/server/src/modules/search/interfaces/search.interface.d.ts
export interface PineconeVectorMetadata extends RecordMetadata {
    patentId: string;
    section?: 'title' | 'abstract' | 'claims' | string;
    ipc: string;
    title?: string;
    abstract?: string;
    claims?: string;
    publicationDate?: string;
    owner?: string;
    assignee?: string;
    country?: string;
}
// ...
export interface PineconeMatchResult {
    id: string;
    score: number;
    metadata?: PineconeVectorMetadata | undefined;
}